refactor forms into separate files for events and places. add pluralize filter to stop review s from happening when reformatting macros.html file.
