+++ what2do2day/reviews/forms.py
@@ -1,4 +1,3 @@
-
 from flask_wtf import FlaskForm
 from wtforms import (StringField,
                      TextAreaField,
@@ -10,34 +9,8 @@ from wtforms.validators import (DataRequired,
                                 Email,
                                 Length, Optional)
 
-
 from what2do2day.templates.validators import validate_rating
-
-
-class RequiredIf(DataRequired):
-    """Validator which makes a field required if another field is set and has a truthy value.
-    from thread in https://gist.github.com/devxoul/7638142
-    Sources:
-        - http://wtforms.simplecodes.com/docs/1.0.1/validators.html
-        - http://stackoverflow.com/questions/8463209/how-to-make-a-field-conditionally-optional-in-wtforms
-        - https://gist.github.com/devxoul/7638142#file-wtf_required_if-py
-    """
-    field_flags = ('requiredif',)
-
-    def __init__(self, message=None, *args, **kwargs):
-        super(RequiredIf).__init__()
-        self.message = message
-        self.conditions = kwargs
-
-    # field is requiring that name field in the form is data value in the form
-    def __call__(self, form, field):
-        for name, data in self.conditions.items():
-            other_field = form[name]
-            if other_field is None:
-                raise Exception('no field named "%s" in form' % name)
-            if other_field.data == data and not field.data:
-                DataRequired.__call__(self, form, field)
-            Optional()(form, field)
+from what2do2day.forms import RequiredIf
 
 
 class ReviewForm(FlaskForm):
@@ -60,4 +33,3 @@ class ReviewForm(FlaskForm):
     def validate_author(self, author):
         if self.has_review.data and str(self.use_place_email.data) == 'n' and len(str(author.data)) < 0:
             raise ValidationError(" you must enter an author")
-
